+++ react-app/src/app/globals.css
@@ -191,9 +191,8 @@ body {
 
 
 .blurred-background {
-    /* filter: blur(5px); /* Adjust the pixel value for more or less blur */
-    /* pointer-events: none; */
-  background-color: rgba(0, 0, 0, 0.4); 
-   z-index: 100;
-   
+    filter: blur(5px); /* Adjust the pixel value for more or less blur */
+    pointer-events: none;
+  /* background-color: rgba(0, 0, 0, 0.4); 
+   z-index: 100; */
 }
