+++ react-app/src/components/Results/ResourceItem.jsx
