+++ react-app/src/components/SearchControl/ClearButton.jsx
